requirements.txt
langchain==0.2.7
langchain-openai==0.1.15
langchain-community==0.2.7
python-dotenv==1.0.1
gradio==4.36.1
google-search-results==2.4.2